@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <Nullable>enable</Nullable>
         <LangVersion>latest</LangVersion>
         <IsPackable>false</IsPackable>
@@ -16,8 +16,8 @@
         <PackageReference Include="HtmlAgilityPack" Version="1.11.71" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
         <PackageReference Include="Moq" Version="4.20.72" />
-        <PackageReference Include="NUnit" Version="4.2.2" />
-        <PackageReference Include="NUnit.Analyzers" Version="4.4.0">
+        <PackageReference Include="NUnit" Version="4.3.1" />
+        <PackageReference Include="NUnit.Analyzers" Version="4.5.0">
           <PrivateAssets>all</PrivateAssets>
           <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
@@ -32,12 +32,17 @@
       <ProjectReference Include="..\Neco.AspNet\Neco.AspNet.csproj" />
       <ProjectReference Include="..\Neco.Common\Neco.Common.csproj" />
       <ProjectReference Include="..\Neco.Search\Neco.Search.csproj" />
+      <ProjectReference Include="..\Neco.Web\Neco.Web.csproj" />
     </ItemGroup>
 
     <ItemGroup>
       <None Update="TestData\**">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </None>
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Web\" />
     </ItemGroup>
 
 </Project>
