@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Net.Http;
 using HttpHandlerConfigurator = System.Action<System.String, System.Net.Http.SocketsHttpHandler, System.Collections.Generic.List<System.IDisposable>>;
+using HttpHandlerDecorator = System.Func<System.String, System.Net.Http.HttpMessageHandler, System.Net.Http.HttpMessageHandler>;
 using HttpClientConfigurator = System.Action<System.String, System.Net.Http.HttpClient>;
 
 public class HttpClientFactoryConfiguration : IEnumerable {
@@ -23,6 +24,13 @@ public class HttpClientFactoryConfiguration : IEnumerable {
 	public ConcurrentDictionary<String, List<HttpHandlerConfigurator>> HttpHandlerConfigurators { get; init; } = new();
 
 	/// <summary>
+	/// Factories to call to decorate generated named handlers. The default name is <see cref="KnownClientNames.Default"/>(<see cref="string.Empty"/>). <br/>
+	/// Predefined decorators are available in <see cref="KnownHttpConfigurators"/>
+	/// </summary>
+	/// <remarks>To apply a decorators to every created handler regardless of name, use <see cref="KnownClientNames.Always"/></remarks>
+	public ConcurrentDictionary<String, List<HttpHandlerDecorator>> HttpHandlerDecorators { get; init; } = new();
+
+	/// <summary>
 	/// Configuration actions to call for generated named clients. The default name is <see cref="KnownClientNames.Default"/>(<see cref="string.Empty"/>). <br/>
 	/// Predefined configurators are available in <see cref="KnownHttpConfigurators"/>
 	/// </summary>
@@ -33,12 +41,12 @@ public class HttpClientFactoryConfiguration : IEnumerable {
 	/// Time after which a handler is marked to be disposed and a new one is created
 	/// </summary>
 	public TimeSpan HandlerLifetime { get; init; } = TimeSpan.FromMinutes(5);
-	
+
 	/// <summary>
 	/// Time between cleanups of handlers that have exceeded their lifetime 
 	/// </summary>
 	public TimeSpan CleanupInterval { get; init; } = TimeSpan.FromMinutes(1);
-	
+
 	/// <summary>
 	/// TRUE (default) to force a garbage collection cycle if at least one handler was disposed during cleanup
 	/// </summary>
@@ -51,19 +59,25 @@ public class HttpClientFactoryConfiguration : IEnumerable {
 	public HttpClientFactoryConfiguration(TimeSpan handlerLifetime) : this() {
 		HandlerLifetime = handlerLifetime;
 	}
-	
+
 	/// <param name="handlerLifetime">Time after which a handler is marked to be disposed and a new one is created</param>
 	/// <param name="cleanupInterval">Time between cleanups of handlers that have exceeded their lifetime </param>
 	/// <param name="garbageCollectAfterDisposedHandler">Time after which a handler is marked to be disposed and a new one is created</param>
-	public HttpClientFactoryConfiguration(TimeSpan handlerLifetime, TimeSpan cleanupInterval, Boolean garbageCollectAfterDisposedHandler) :this(handlerLifetime) {
+	public HttpClientFactoryConfiguration(TimeSpan handlerLifetime, TimeSpan cleanupInterval, Boolean garbageCollectAfterDisposedHandler) : this(handlerLifetime) {
 		CleanupInterval = cleanupInterval;
 		GarbageCollectAfterDisposedHandler = garbageCollectAfterDisposedHandler;
 	}
-	
+
 	public void Add(String name, HttpHandlerConfigurator configurator) {
 		HttpHandlerConfigurators
 			.GetOrAdd(name, _ => [])
 			.Add(configurator);
+	}
+
+	public void Add(String name, HttpHandlerDecorator decorator) {
+		HttpHandlerDecorators
+			.GetOrAdd(name, _ => [])
+			.Add(decorator);
 	}
 
 	public void Add(String name, HttpClientConfigurator configurator) {
@@ -80,6 +94,14 @@ public class HttpClientFactoryConfiguration : IEnumerable {
 		Add(tpl.name ?? KnownClientNames.Default, (_, handler, _) => tpl.configurator(handler));
 	}
 
+	public void Add((String? name, HttpHandlerDecorator decorator) tpl) {
+		Add(tpl.name ?? KnownClientNames.Default, tpl.decorator);
+	}
+
+	public void Add((String? name, Func<HttpMessageHandler, HttpMessageHandler> decorator) tpl) {
+		Add(tpl.name ?? KnownClientNames.Default, (_, handler) => tpl.decorator(handler));
+	}
+
 	public void Add((String? name, HttpClientConfigurator configurator) tpl) {
 		Add(tpl.name ?? KnownClientNames.Default, tpl.configurator);
 	}
@@ -90,7 +112,10 @@ public class HttpClientFactoryConfiguration : IEnumerable {
 
 	public void Add(HttpHandlerConfigurator configurator) => Add((KnownClientNames.Default, configurator));
 	public void Add(Action<SocketsHttpHandler> configurator) => Add((KnownClientNames.Default, configurator));
-	
+
+	public void Add(HttpHandlerDecorator decorator) => Add((KnownClientNames.Default, decorator));
+	public void Add(Func<HttpMessageHandler, HttpMessageHandler> decorator) => Add((KnownClientNames.Default, decorator));
+
 	public void Add(HttpClientConfigurator configurator) => Add((KnownClientNames.Default, configurator));
 	public void Add(Action<HttpClient> configurator) => Add((KnownClientNames.Default, configurator));
 
@@ -101,4 +126,44 @@ public class HttpClientFactoryConfiguration : IEnumerable {
 	public IEnumerator GetEnumerator() => throw new InvalidOperationException();
 
 	#endregion
+
+	public void Add(HttpClientFactoryConfiguration configuration) {
+		foreach ((String? name, List<HttpHandlerConfigurator>? configurators) in configuration.HttpHandlerConfigurators) {
+			foreach (HttpHandlerConfigurator configurator in configurators) {
+				Add(name, configurator);
+			}
+		}
+		
+		foreach ((String? name, List<HttpHandlerDecorator>? decorators) in configuration.HttpHandlerDecorators) {
+			foreach (HttpHandlerDecorator decorator in decorators) {
+				Add(name, decorator);
+			}
+		}
+		
+		foreach ((String? name, List<HttpClientConfigurator>? configurators) in configuration.HttpClientConfigurators) {
+			foreach (HttpClientConfigurator configurator in configurators) {
+				Add(name, configurator);
+			}
+		}
+	}
+
+	public void Add((string name, HttpClientFactoryConfiguration configuration) tpl) {
+		foreach ((_, List<HttpHandlerConfigurator>? configurators) in tpl.configuration.HttpHandlerConfigurators) {
+			foreach (HttpHandlerConfigurator configurator in configurators) {
+				Add(tpl.name, configurator);
+			}
+		}
+		
+		foreach ((_, List<HttpHandlerDecorator>? decorators) in tpl.configuration.HttpHandlerDecorators) {
+			foreach (HttpHandlerDecorator decorator in decorators) {
+				Add(tpl.name, decorator);
+			}
+		}
+		
+		foreach ((_, List<HttpClientConfigurator>? configurators) in tpl.configuration.HttpClientConfigurators) {
+			foreach (HttpClientConfigurator configurator in configurators) {
+				Add(tpl.name, configurator);
+			}
+		}
+	}
 };
